+++ packages/lwc-compiler/src/bundler/bundler.ts
@@ -1,9 +1,9 @@
 import { rollup } from "rollup";
-import rollupPluginReplace from "rollup-plugin-replace";
 
 import { MetadataCollector, BundleMetadata } from "./meta-collector";
 import rollupModuleResolver from "../rollup-plugins/module-resolver";
 
+import rollupEnvReplacement from "../rollup-plugins/env-replacement";
 import rollupTransform from "../rollup-plugins/transform";
 import rollupCompat from "../rollup-plugins/compat";
 import rollupMinify from "../rollup-plugins/minify";
@@ -49,37 +49,44 @@ function handleRollupWarning(diagnostics: Diagnostic[]) {
 }
 
 export async function bundle(
-    options: NormalizedCompilerOptions
+    options: NormalizedCompilerOptions,
 ): Promise<BundleReport> {
     validateNormalizedOptions(options);
 
     const { outputConfig, name, namespace } = options;
+    const { sourcemap } = outputConfig;
 
     // TODO: remove format option once tests are converted to 'amd' format
     const format = (outputConfig as any).format || DEFAULT_FORMAT;
-    const { sourcemap } = outputConfig;
 
     const diagnostics: Diagnostic[] = [];
-
     const metadataCollector = new MetadataCollector();
 
-    const plugins = [
-        rollupPluginReplace({
-            "process.env.NODE_ENV": JSON.stringify(outputConfig.env.NODE_ENV),
-            sourcemap
-        }),
+    const plugins: any[] = [
         rollupModuleResolver({
-            metadataCollector,
-            options
+            options,
         }),
-        rollupTransform({
-            metadataCollector,
-            options
-        })
     ];
 
+    // Run environment variable replacement first. This ensures that the source code is still untouched
+    // at this point.
+    if (Object.keys(outputConfig.env).length) {
+        plugins.push(
+            rollupEnvReplacement({
+                options,
+            }),
+        );
+    }
+
+    plugins.push(
+        rollupTransform({
+            metadataCollector,
+            options,
+        }),
+    );
+
     if (outputConfig.compat) {
-        plugins.push(rollupCompat(outputConfig.resolveProxyCompat, sourcemap));
+        plugins.push(rollupCompat(sourcemap));
     }
 
     if (outputConfig.minify) {