+++ packages/lwc-compiler/src/bundler/meta-collector.ts
@@ -6,11 +6,18 @@ import {
     ClassMember
 } from "babel-plugin-transform-lwc-class";
 
+import { TemplateModuleDependency } from "lwc-template-compiler";
+
 import { ModuleImportLocation } from "./import-location-collector";
 
 export type MetadataDecorators = Array<
     ApiDecorator | TrackDecorator | WireDecorator
 >;
+
+export interface TemplateModuleDependencies {
+    templatePath: string;
+    moduleDependencies: TemplateModuleDependency[];
+}
 
 export interface BundleMetadata {
     decorators: MetadataDecorators;
@@ -18,6 +25,7 @@ export interface BundleMetadata {
     classMembers: ClassMember[];
     declarationLoc?: Location;
     doc?: string;
+    experimentalTemplateDependencies?: TemplateModuleDependencies[];
 }
 
 export class MetadataCollector {
@@ -25,6 +33,7 @@ export class MetadataCollector {
         ApiDecorator | TrackDecorator | WireDecorator
     > = [];
     private importLocations: ModuleImportLocation[] = [];
+    private experimentalTemplateDependencies?: TemplateModuleDependencies[];
     private classMembers: ClassMember[] = [];
     private declarationLoc?: Location;
     private doc?: string;
@@ -51,6 +60,17 @@ export class MetadataCollector {
         this.doc = doc;
     }
 
+    public collectExperimentalTemplateDependencies(
+        templatePath: string, templateDependencies: TemplateModuleDependency[]) {
+        if (!this.experimentalTemplateDependencies) {
+            this.experimentalTemplateDependencies = [];
+        }
+        this.experimentalTemplateDependencies.push({
+            templatePath,
+            moduleDependencies: templateDependencies
+        });
+    }
+
     public getMetadata(): BundleMetadata {
         return {
             decorators: this.decorators,
@@ -58,6 +78,7 @@ export class MetadataCollector {
             classMembers: this.classMembers,
             declarationLoc: this.declarationLoc,
             doc: this.doc,
+            experimentalTemplateDependencies: this.experimentalTemplateDependencies
         };
     }
 }