+++ packages/lwc-compiler/src/rollup-plugins/compat.ts
@@ -1,15 +1,21 @@
 import * as babel from '@babel/core';
 import * as presetCompat from 'babel-preset-compat';
-import { BABEL_CONFIG_BASE } from '../babel-plugins';
-import { OutputProxyCompatConfig } from "../compiler/options";
 
-export default function(proxyCompatOption: OutputProxyCompatConfig | undefined, sourcemaps?: boolean | 'inline') {
-    const overrides = sourcemaps !== undefined ? { sourceMaps: sourcemaps } : {};
-    const config = Object.assign({}, BABEL_CONFIG_BASE, overrides, {
-        presets: [
-            [presetCompat, { proxy: true, resolveProxyCompat: proxyCompatOption || { independent: "proxy-compat" } }],
-        ],
-    });
+import { BABEL_CONFIG_BASE } from '../babel-plugins';
+
+const BABEL_CONFIG_CONFIG = {
+    ...BABEL_CONFIG_BASE,
+    presets: [
+        [presetCompat, { proxy: true }],
+    ],
+};
+
+export default function(sourcemaps?: boolean) {
+    const config = Object.assign(
+        {},
+        BABEL_CONFIG_CONFIG,
+        sourcemaps !== undefined ? { sourceMaps: sourcemaps } : {}
+    );
 
     return {
         name: "lwc-compat",