+++ packages/lwc-engine/src/framework/__tests__/vm.spec.ts
@@ -1,7 +1,7 @@
+import { compileTemplate } from 'test-utils';
 import { createElement, LightningElement } from '../main';
 import { ViewModelReflection } from "../utils";
 import { getErrorComponentStack } from "../vm";
-import { compileTemplate } from 'test-utils';
 
 describe('vm', () => {
     describe('insertion index', () => {
@@ -28,15 +28,21 @@ describe('vm', () => {
 
         it('should assign bigger idx to children', () => {
             let vm1: VM, vm2: VM;
+
             class ChildComponent4 extends LightningElement {
                 constructor() {
                     super();
                     vm2 = this[ViewModelReflection];
                 }
             }
-            function html($api) {
-                return [$api.c('x-bar', ChildComponent4, {})];
-            }
+
+            const html  = compileTemplate(`
+                <template>
+                    <x-bar></x-bar>
+                </template>
+            `, {
+                modules: { 'x-bar': ChildComponent4 }
+            });
             class MyComponent4 extends LightningElement {
                 constructor() {
                     super();
@@ -46,6 +52,7 @@ describe('vm', () => {
                     return html;
                 }
             }
+
             const elm = createElement('x-foo', { is: MyComponent4 });
             document.body.appendChild(elm);
             expect(vm1.idx).toBeGreaterThan(0);
@@ -65,9 +72,14 @@ describe('vm', () => {
                     counter++;
                 }
             }
-            function html($api) {
-                return [$api.c('x-bar', ChildComponent5, {})];
-            }
+
+            const html  = compileTemplate(`
+                <template>
+                    <x-bar></x-bar>
+                </template>
+            `, {
+                modules: { 'x-bar': ChildComponent5 }
+            });
             class MyComponent5 extends LightningElement {
                 constructor() {
                     super();
@@ -77,6 +89,7 @@ describe('vm', () => {
                     return html;
                 }
             }
+
             const elm = createElement('x-foo', { is: MyComponent5 });
             document.body.appendChild(elm);
             expect(vm1.idx).toBeGreaterThan(0);
@@ -103,9 +116,12 @@ describe('vm', () => {
                     vm = this[ViewModelReflection];
                 }
             }
-            const html  = compileTemplate(
-                `<template><x-child></x-child></template>`,
-                { modules: { 'x-child': ChildComponentCs }
+            const html  = compileTemplate(`
+                <template>
+                    <x-child></x-child>
+                </template>
+            `, {
+                modules: { 'x-child': ChildComponentCs }
             });
             class ParentComponentCs extends LightningElement {
                 constructor() {