+++ packages/lwc-template-compiler/src/index.ts
@@ -6,7 +6,11 @@ import generate from './codegen';
 
 import { TEMPLATE_MODULES_PARAMETER } from './shared/constants';
 import { CompilationMetadata, CompilationWarning } from './shared/types';
-import { kebabcaseToCamelcase } from './shared/naming';
+
+export {
+    ModuleDependency as TemplateModuleDependency,
+    DependencyParameter as TemplateModuleDependencyParameter
+} from './shared/types';
 
 export default function compiler(
     source: string,
@@ -49,7 +53,7 @@ export default function compiler(
         metadata: {
             definedSlots: state.slots,
             templateUsedIds: state.ids,
-            templateDependencies: state.dependencies.map(kebabcaseToCamelcase),
+            templateDependencies: state.extendedDependencies,
         },
     };
 }