+++ scripts/jest/root.config.js
@@ -1,5 +1,6 @@
 module.exports = {
     rootDir: '../..',
+
     projects: [
         '<rootDir>/packages/babel-plugin-transform-lwc-class',
         '<rootDir>/packages/lwc-compiler',